Move UsedDays and UsedDaysOverview to the overview package
