+++ src/main/java/org/synyx/urlaubsverwaltung/overview/OverviewViewController.java
@@ -25,7 +25,6 @@ import org.synyx.urlaubsverwaltung.settings.SettingsService;
 import org.synyx.urlaubsverwaltung.sicknote.SickNote;
 import org.synyx.urlaubsverwaltung.sicknote.SickNoteService;
 import org.synyx.urlaubsverwaltung.sicknote.web.ExtendedSickNote;
-import org.synyx.urlaubsverwaltung.statistics.web.UsedDaysOverview;
 import org.synyx.urlaubsverwaltung.util.DateUtil;
 import org.synyx.urlaubsverwaltung.workingtime.WorkDaysCountService;
